refactor set #4

+++ crypto/threads_win.c
@@ -305,7 +305,7 @@ int CRYPTO_THREAD_INTERN_join(void * thread, unsigned long * retval)
 
 void CRYPTO_THREAD_INTERN_exit(unsigned long retval)
 {
-    ExitThread(retval);
+    ExitThread((DWORD) retval);
 }
 
 CRYPTO_RWLOCK *CRYPTO_THREAD_lock_new(void)
